SPA - migrate login-error page
